Remove testpypi and github release workflow.
TestPyPi seems to not work when commiting the same version twice.

The github release workflow created a new release. But since this workflow is triggered by the release, we don't want that. Hence we stop after signing. The signing action seems to automatically upload all artifacts + signatures, when a matching release was found.